requirements.txt
evdev==1.5.0
PyMySQL==1.0.2
RPi.GPIO==0.7.1
SQLAlchemy==1.4.39
